lib: Move the reusable service code into lib directory
And bundle it in the way that webpack will. This lets us migrate
towards the bundling we're starting to use in other pull requests.
