+++ pkg/playground/service.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="../base1/cockpit.css" type="text/css" rel="stylesheet">
     <script src="../base1/bundle.js"></script>
-    <script src="service.js"></script>
+    <script src="jquery-service.js"></script>
 </head>
 <body hidden>
   <table>
