+++ pkg/storaged/service.js
